+++ webapp/src/components/App/Singlestudy/explore/Modelization/Areas/Hydro/utils.ts
@@ -37,6 +37,7 @@ export interface HydroMatrixProps {
   fetchFn?: fetchMatrixFn;
   disableEdit?: boolean;
   enablePercentDisplay?: boolean;
+  stretch?: boolean; // TODO: Remove this once the `EditableMatrix` component is refactored
 }
 
 type Matrices = Record<HydroMatrixType, HydroMatrixProps>;
@@ -125,6 +126,7 @@ export const MATRICES: Matrices = {
       "Pumping Max Energy (Hours at Pmax)",
     ],
     stats: MatrixStats.NOCOL,
+    stretch: false,
   },
   [HydroMatrixType.ReservoirLevels]: {
     title: "Reservoir Levels",
@@ -185,6 +187,7 @@ export const MATRICES: Matrices = {
       "December",
     ],
     stats: MatrixStats.NOCOL,
+    stretch: false,
   },
   [HydroMatrixType.Allocation]: {
     title: "Allocation",