+++ webapp/src/components/App/Singlestudy/explore/Results/ResultDetails/index.tsx
@@ -49,6 +49,7 @@ import ButtonBack from "../../../../../common/ButtonBack";
 import BooleanFE from "../../../../../common/fieldEditors/BooleanFE";
 import SelectFE from "../../../../../common/fieldEditors/SelectFE";
 import NumberFE from "../../../../../common/fieldEditors/NumberFE";
+import moment from "moment";
 
 function ResultDetails() {
   const { study } = useOutletContext<{ study: StudyMetadata }>();
@@ -148,15 +149,27 @@ function ResultDetails() {
     },
   );
 
+  // !NOTE: Workaround to display the date in the correct format, to be replaced by a proper solution.
+  const dateTimeFromIndex = useMemo(() => {
+    if (!matrixRes.data) return [];
+
+    return matrixRes.data.index.map((dateTime) => {
+      const parsedDate = moment(dateTime, "MM/DD HH:mm");
+      return parsedDate.format("ddd D MMM HH:mm");
+    });
+  }, [matrixRes.data]);
+
   ////////////////////////////////////////////////////////////////
   // Event Handlers
   ////////////////////////////////////////////////////////////////
 
   const handleItemTypeChange: ToggleButtonGroupProps["onChange"] = (
     _,
-    value: OutputItemType,
+    newValue: OutputItemType,
   ) => {
-    setItemType(value);
+    if (newValue && newValue !== itemType) {
+      setItemType(newValue);
+    }
   };
 
   const handleDownload = (matrixData: MatrixType, fileName: string): void => {
@@ -366,6 +379,8 @@ function ResultDetails() {
                     <EditableMatrix
                       matrix={matrix}
                       matrixTime={false}
+                      rowNames={dateTimeFromIndex}
+                      stretch={false}
                       readOnly
                     />
                   )